+++ database.php
@@ -1,5 +1,5 @@
 <?php
-
+require_once 'functions.php';
 
 $db = new PDO ('mysql:host=db; dbname=mohammad-collection', 'root', 'password');
 $db->setAttribute(PDO::ATTR_DEFAULT_FETCH_MODE, PDO::FETCH_ASSOC);
@@ -8,19 +8,5 @@ $query = $db->prepare(" SELECT `artist`, `year-made`, `painting-name`,`image-lin
 
 $query->execute();
 $results = $query->fetchAll();
-
-echo '<pre>';
-var_dump($results);
-echo '</pre>';
-
-function displayDB(array $results) : string {
-    $dbResult = ' ';
-    foreach($results as $result){
-       $dbResult .= '<div>'. '<img src="images/' . $result['image-link'] . '" alt="a painting of  '
-           . $result['painting-name'] .' " /><p>'
-           . $result['artist'] . '<br>' . $result['painting-name'] . '<br>' . $result['year-made'] .  '</p></div>';
-    }
-    return $dbResult;
-}
 
 $displayResults = displayDB($results);
