+++ index.php
@@ -1,6 +1,7 @@
 <?php
 
 require_once 'database.php';
+require_once 'functions.php';
 ?>
 
 <!DOCTYPE html>
